Move platform-specific code under build tags; fix corresponding tests
