+++ ffmpeg_test.go
@@ -277,13 +277,6 @@ func TestCompile(t *testing.T) {
 	assert.Equal(t, out.Compile().Args, []string{"ffmpeg", "-i", "dummy.mp4", "dummy2.mp4"})
 }
 
-func TestCompileWithOptions(t *testing.T) {
-	out := Input("dummy.mp4").Output("dummy2.mp4")
-	cmd := out.Compile(SeparateProcessGroup())
-	assert.Equal(t, cmd.SysProcAttr.Pgid, 0)
-	assert.True(t, cmd.SysProcAttr.Setpgid)
-}
-
 func TestPipe(t *testing.T) {
 
 	width, height := 32, 32
